+++ src/gamspy/_convert.py
@@ -291,7 +291,7 @@ class GamsConverter:
             )
 
         # 2. Load the data from gdx
-        load_str = f"$gdxLoadAll {os.path.abspath(self.gdx_path)}"
+        load_str = f"$onMultiR\n$gdxLoadAll {os.path.abspath(self.gdx_path)}\n$$offMulti"
 
         # 3. Definitions
         definitions = self.get_definitions()