+++ src/gamspy/_symbols/equation.py
@@ -346,6 +346,7 @@ class Equation(gt.Equation, Symbol):
             if records is not None:
                 self.setRecords(records, uels_on_axes=uels_on_axes)
             else:
+                self.modified = False
                 self.container._synch_with_gams()
 
             container._options.miro_protect = previous_state
@@ -1140,6 +1141,9 @@ class Equation(gt.Equation, Symbol):
 
         if self.description:
             output += ' "' + self.description + '"'
+
+        if self.records is None:
+            output += " / /"
 
         output += ";"
         return output