+++ tests/unit/test_parameter.py
@@ -89,7 +89,7 @@ def test_parameter_string(data):
     assert a.getDeclaration() == 'Parameter a(i) "capacities";'
 
     b = Parameter(m, "b")
-    assert b.getDeclaration() == "Parameter b;"
+    assert b.getDeclaration() == "Parameter b / /;"
     assert (b == 5).gamsRepr() == "(b eq 5)"
     assert (-b).getDeclaration() == "( - b)"
     assert (b != 5).gamsRepr() == "(b ne 5)"
@@ -120,7 +120,7 @@ def test_implicit_parameter_string(data):
     A = Parameter(cont, "A", domain=[s, m])
 
     A.domain = ["s", "m"]
-    assert A.getDeclaration() == "Parameter A(*,*);"
+    assert A.getDeclaration() == "Parameter A(*,*) / /;"
 
 
 def test_parameter_assignment(data):