+++ CHANGELOG.md
@@ -18,6 +18,7 @@ GAMSPy 1.9.0
   - Do not validate equation definitions in case the container was restarted from a save file (.g00 file).
   - Propagate the output option of the container to `model.freeze`.
   - Raise warning in case the frozen solve is interrupted.
+  - Improve the performance of symbol declarations without any records.
 - Documentation
   - Add additional instructions to deploy a GAMSPy/GAMS MIRO model.
   - Fix name mismatch between the argument name and the docstring of loadRecordsFromGdx function.